autopep8

+++ malaya_speech/train/model/demucs/layer.py
@@ -3,23 +3,25 @@ from tensorflow.python.ops.init_ops import Initializer
 from ..utils import shape_list
 
 # check Table 4, https://arxiv.org/pdf/1911.13254.pdf
+
+
 class ConvScaling(Initializer):
     def __init__(
-        self, scale = 1.0, reference = 0.1, seed = None, dtype = tf.float32
+        self, scale=1.0, reference=0.1, seed=None, dtype=tf.float32
     ):
         self.scale = scale
         self.reference = reference
         self.seed = seed
         self.dtype = dtype
 
-    def __call__(self, shape, dtype = None, partition_info = None):
+    def __call__(self, shape, dtype=None, partition_info=None):
         stdv = 1.0 / (shape[0] * shape[1])
         w = tf.random.uniform(
             shape,
-            minval = -stdv,
-            maxval = stdv,
-            dtype = self.dtype,
-            seed = self.seed,
+            minval=-stdv,
+            maxval=stdv,
+            dtype=self.dtype,
+            seed=self.seed,
         )
         std = tf.math.reduce_std(w)
         scale = (std / self.reference) ** 0.5
@@ -34,14 +36,14 @@ class ConvScaling(Initializer):
         }
 
 
-def glu(x, dims = 2, kernel_initializer = ConvScaling):
+def glu(x, dims=2, kernel_initializer=ConvScaling):
     def conv(x):
         return tf.layers.conv1d(
             x,
             x.shape[-1],
             1,
-            padding = 'same',
-            kernel_initializer = kernel_initializer,
+            padding='same',
+            kernel_initializer=kernel_initializer,
         )
 
     splitted = tf.split(x, 2, dims)
@@ -60,12 +62,12 @@ def center_trim(tensor, reference):
             f'tensor must be larger than reference. Delta is {delta}.'
         )
     if delta:
-        tensor = tensor[:, delta // 2 : -(delta - delta // 2)]
+        tensor = tensor[:, delta // 2: -(delta - delta // 2)]
     return tensor
 
 
 class BLSTM:
-    def __init__(self, dim, num_layers = 2):
+    def __init__(self, dim, num_layers=2):
         self.dim = dim
         self.num_layers = num_layers
 
@@ -73,7 +75,7 @@ class BLSTM:
         for _ in range(num_layers):
             self.lstm.add(
                 tf.keras.layers.Bidirectional(
-                    tf.keras.layers.LSTM(dim, return_sequences = True)
+                    tf.keras.layers.LSTM(dim, return_sequences=True)
                 )
             )
         self.dense = tf.keras.layers.Dense(dim)
@@ -90,18 +92,18 @@ class Conv1DTranspose(tf.keras.layers.Layer):
         kernel_size,
         strides,
         activation,
-        kernel_initializer = ConvScaling,
+        kernel_initializer=ConvScaling,
         **kwargs,
     ):
         super(Conv1DTranspose, self).__init__(
-            name = 'Conv1DTranspose', **kwargs
+            name='Conv1DTranspose', **kwargs
         )
         self.conv = tf.keras.layers.Conv2DTranspose(
             filters,
             (kernel_size, 1),
-            strides = (strides, 1),
-            activation = activation,
-            kernel_initializer = kernel_initializer,
+            strides=(strides, 1),
+            activation=activation,
+            kernel_initializer=kernel_initializer,
         )
 
     def call(self, x):
